+++ bindings/java/java-xconfig/xconfig-base/src/main/java/com/tuenti/xconfig/BreedXConfig.java
@@ -3,7 +3,6 @@ package com.tuenti.xconfig;
 import static com.tuenti.xconfig.utils.StringUtils.split;
 
 import com.tuenti.xconfig.exception.XConfigKeyNotFoundException;
-import com.tuenti.xconfig.exception.XConfigWrongTypeCastingException;
 import com.tuenti.xconfig.type.XConfigMap;
 import com.tuenti.xconfig.type.XConfigValue;
 
@@ -16,16 +15,16 @@ import com.tuenti.xconfig.type.XConfigValue;
  * will also be generated using all the combinations of parent/child keys (see BreedsCombiner).
  */
 public class BreedXConfig extends XConfigBase {
+	private static final BreedsCombiner BREEDS_COMBINER = new BreedsCombiner();
 
-	private XConfig xconfig;
-	private String[][] breed;
-	private BreedsCombiner breedsCombiner = new BreedsCombiner();
+	final private XConfig xconfig;
+	final private String[][] breed;
 
 	public BreedXConfig(XConfig xconfig, String[][] breedsArray) {
 		if (xconfig instanceof BreedXConfig) {
 			BreedXConfig breedBaseConfig = (BreedXConfig) xconfig;
 			this.xconfig = breedBaseConfig.xconfig;
-			this.breed = breedsCombiner.combineBreeds(breedBaseConfig.breed, breedsArray);
+			this.breed = BREEDS_COMBINER.combineBreeds(breedBaseConfig.breed, breedsArray);
 			
 		} else {
 			this.xconfig = xconfig;
@@ -45,30 +44,30 @@ public class BreedXConfig extends XConfigBase {
 
 	@Override
 	public XConfigValue getValue(String key) throws XConfigKeyNotFoundException {
-		XConfigValue value = null;
-		boolean isMap = true;
-		boolean found = false;
+		XConfigMap mergedMap = null;
 
 		String[] breedKeys = getBreedKeys(key);
 
 		// Iterate the keys in reverse order
-		// We only continue iterating if we received a map
-		for (int i=breedKeys.length-1; i>=0 && isMap; i--) {
+		// We only continue iterating if we received a map or the key doesn't exists
+		for (int i=breedKeys.length-1; i>=0; i--) {
 			String breedKey = breedKeys[i];
 
 			try {
 				XConfigValue newValue = xconfig.getValue(breedKey);
-				isMap = newValue instanceof XConfigMap;
-				value = found ? calculateOverride(newValue, value) : newValue;
-				found = true;
-			} catch (XConfigKeyNotFoundException e) { }
+				if (newValue instanceof XConfigMap) {
+					XConfigMap newValueAsMap = (XConfigMap) newValue;
+					mergedMap = (mergedMap != null) ? newValueAsMap.mergeWith(mergedMap) : newValueAsMap;
+				} else {
+					return (mergedMap != null) ? mergedMap : newValue;
+				}
+			} catch (XConfigKeyNotFoundException ignored) { }
 		}
 
-		if (!found) {
+		if (mergedMap == null) {
 			throw new XConfigKeyNotFoundException(key);
 		}
-
-		return value;
+		return mergedMap;
 	}
 
 	@Override
@@ -91,8 +90,6 @@ public class BreedXConfig extends XConfigBase {
 
 	/**
 	 * Get all the keys we will be trying when getting a specified breed
-	 * @param originalKey
-	 * @return
 	 */
 	private String[] getBreedKeys(String originalKey) {
 		int keyCount = breed.length + 1;
@@ -112,20 +109,5 @@ public class BreedXConfig extends XConfigBase {
 			keys[i+1] = newKey;
 		}
 		return keys;
-	}
-
-	private XConfigValue calculateOverride(XConfigValue baseValue, XConfigValue overrideValue) {
-		if (baseValue != null && overrideValue != null) {
-			try {
-				XConfigMap baseMap = baseValue.getAsMap();
-				XConfigMap overrideMap = overrideValue.getAsMap();
-
-				if (baseMap != null && overrideMap != null) {
-					baseMap.overrideWith(overrideMap);
-					return baseMap;
-				}
-			} catch (XConfigWrongTypeCastingException e) { }
-		}
-		return overrideValue;
 	}
 }