+++ bindings/java/java-xconfig/xconfig-base/src/test/java/unit/com/tuenti/xconfig/BreedXConfigTest.java
@@ -2,11 +2,21 @@ package com.tuenti.xconfig;
 
 import static org.junit.Assert.*;
 
+import com.tuenti.xconfig.cache.XConfigCache;
+import com.tuenti.xconfig.cache.XConfigCacheHook;
+import com.tuenti.xconfig.cache.XConfigCacheProvider;
+import com.tuenti.xconfig.cache.XConfigCacheWrapper;
+import com.tuenti.xconfig.type.XConfigInteger;
+import com.tuenti.xconfig.type.XConfigMap;
+import com.tuenti.xconfig.type.XConfigValue;
 import org.junit.Test;
 
 import com.tuenti.xconfig.exception.XConfigKeyNotFoundException;
 import com.tuenti.xconfig.testutils.XConfigBaseTestable;
 import com.tuenti.xconfig.type.XConfigString;
+
+import java.util.HashMap;
+import java.util.Map;
 
 public class BreedXConfigTest {
 	
@@ -215,7 +225,7 @@ public class BreedXConfigTest {
 		BreedXConfig breedConfig = new BreedXConfig(base, BREEDS);
 		assertEquals(new XConfigString("value"), breedConfig.getValue("root/key"));
 	}
-	
+
 	@Test(expected=XConfigKeyNotFoundException.class)
 	public void testGetValueForBreededBaseXConfigAndMissingValue() {
 		XConfig base = breededConfig(
@@ -226,13 +236,88 @@ public class BreedXConfigTest {
 		BreedXConfig breedConfig = new BreedXConfig(base, BREEDS);
 		breedConfig.getValue("root/key2");
 	}
-	
-	
+
+	@Test
+	public void testGetValueBreededFromScalarToMap() {
+		XConfigBaseTestable base = new XConfigBaseTestable()
+				.addKey("root/key", newXConfigMap("a", "a-value"))
+				.addKey("root_breeds/breed0/breed0Value/key", new XConfigInteger(9));
+		BreedXConfig breedXConfig = new BreedXConfig(base, BASE_BREEDS);
+
+		XConfigValue value = breedXConfig.getValue("root/key");
+		assertEquals(new XConfigInteger(9), value);
+	}
+
+	@Test
+	public void testGetValueBreededFromMapToScalar() {
+		XConfigBaseTestable base = new XConfigBaseTestable()
+				.addKey("root/key", new XConfigInteger(9))
+				.addKey("root_breeds/breed0/breed0Value/key", newXConfigMap("a", "a-value"));
+		BreedXConfig breedXConfig = new BreedXConfig(base, BASE_BREEDS);
+
+		XConfigValue value = breedXConfig.getValue("root/key");
+		assertEquals(newXConfigMap("a", "a-value"), value);
+	}
+
+	/** Test for bug #FRWK-532 */
+	@Test
+	public void testCachedXConfigNotAffectedByBreedOverride() {
+		XConfigBaseTestable xconfig = new XConfigBaseTestable()
+				.addKey("base/mapValue", newXConfigMap(
+						"a", "a-value",
+						"b", "b-value"
+				)).addKey("base_breeds/breed0/breed0Value/mapValue", newXConfigMap(
+						"a", "overridden-a",
+						"c", "c-value"
+				));
+		XConfig cachedXConfig = cacheXConfig(xconfig);
+		BreedXConfig breedXConfig = new BreedXConfig(cachedXConfig, BASE_BREEDS);
+
+		XConfigMap baseMap = cachedXConfig.getAsMap("base/mapValue");
+		XConfigMap overriddenMap = breedXConfig.getAsMap("base/mapValue");
+		XConfigMap baseMapReread = cachedXConfig.getAsMap("base/mapValue");
+		XConfigMap overriddenMapReread = breedXConfig.getAsMap("base/mapValue");
+
+		XConfigMap expectedBaseMap = newXConfigMap(
+				"a", "a-value",
+				"b", "b-value"
+		);
+		assertEquals(expectedBaseMap, baseMap);
+		assertEquals(expectedBaseMap, baseMapReread);
+
+		XConfigMap expectedOverriddenMap = newXConfigMap(
+				"a", "overridden-a",
+				"b", "b-value",
+				"c", "c-value"
+		);
+		assertEquals(expectedOverriddenMap, overriddenMap);
+		assertEquals(expectedOverriddenMap, overriddenMapReread);
+	}
+
 	private XConfig xconfig(String ... keys) {
 		return new XConfigBaseTestable(keys);
 	}
-	
+
 	private BreedXConfig breededConfig(String ... keys) {
 		return new BreedXConfig(xconfig(keys), BASE_BREEDS);
+	}
+
+	private XConfigMap newXConfigMap(String... keyValueInPairs) {
+		Map<String, XConfigValue> sourceMap = new HashMap<>();
+		for (int i=0; i<keyValueInPairs.length-1; i+=2) {
+			sourceMap.put(keyValueInPairs[i], new XConfigString(keyValueInPairs[i + 1]));
+		}
+		return new XConfigMap(sourceMap);
+	}
+
+	private XConfig cacheXConfig(XConfig base) {
+		return new XConfigCacheWrapper(base, (config, currentCache) -> {
+			if (currentCache == null) {
+				XConfigCache xConfigCache = new XConfigCache(0, new XConfigCacheHook() {});
+				return new XConfigCacheProvider.ReloadResult(true, xConfigCache);
+			} else {
+				return new XConfigCacheProvider.ReloadResult(false, currentCache);
+			}
+		});
 	}
 }