+++ bindings/java/java-xconfig/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Java XConfig Changelog
 
+### Next
+ - [FRWK-532](https://jira.tid.es/browse/FRWK-532) Fixed BreedXConfig mutating the cached map when merging breeds.
+
 ### 1.4.1
  - [FRWK-511](https://jira.tid.es/browse/FRWK-511) Changed the xconfig cache to a bounded implementation to mitigate memory pressure when a service reads a lot of different keys.
     * Current implementation uses a bounded cache with lru-like evictions.